refactor#60 maps component file location
