+++ core/designsystem/src/main/java/com/teamsolply/solply/designsystem/component/dialog/SolplyBasicDialog.kt
@@ -1,0 +1,25 @@
+package com.teamsolply.solply.designsystem.component.dialog
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.window.Dialog
+import androidx.compose.ui.window.DialogProperties
+
+@Composable
+fun SolplyBasicDialog(
+    onDismissRequest: () -> Unit,
+    content: @Composable () -> Unit
+) {
+    Dialog(
+        onDismissRequest = onDismissRequest,
+        properties = DialogProperties(dismissOnClickOutside = false, dismissOnBackPress = true)
+    ) {
+        Column(
+            modifier = Modifier.fillMaxSize()
+        ) {
+            content.invoke()
+        }
+    }
+}
