+++ feature/maps/src/main/java/com/teamsolply/solply/maps/component/bottomsheet/PlaceDetailBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.teamsolply.solply.maps.placedetail
+package com.teamsolply.solply.maps.component.bottomsheet
 
 import ClickableAnnotatedText
 import androidx.compose.foundation.Image
