+++ feature/maps/src/main/java/com/teamsolply/solply/maps/component/dialog/CourseSaveBottomSheet.kt
@@ -1,0 +1,7 @@
+package com.teamsolply.solply.maps.component.dialog
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun CourseSaveBottomSheet() {
+}
