+++ feature/maps/src/main/java/com/teamsolply/solply/maps/util/DragDropExtension.kt
@@ -1,4 +1,4 @@
-package com.teamsolply.solply.maps.editcourse.extension
+package com.teamsolply.solply.maps.util
 
 import android.annotation.SuppressLint
 import androidx.compose.animation.core.FastOutSlowInEasing
@@ -11,7 +11,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.zIndex
-import com.teamsolply.solply.maps.editcourse.interaction.DragDropState
 import kotlinx.collections.immutable.PersistentList
 
 internal inline fun <T : Any> LazyListScope.draggableItems(
